% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno_info}
\alias{anno_info}
\title{Annotation information}
\format{A data frame with 1236881 rows and 9 variables}
\usage{
data(anno_info)
}
\description{
A dataset containing the annotation information from v2.1 of the Phaseolus
    vulgaris genome. The variables are as follows:
}
\details{
\itemize{
  \item pacId.
  \item locusName. Locus name in Phaseolus vulgaris.
  \item transcriptName.
  \item peptideName.
  \item Pfam.
  \item Panther.
  \item KOG.
  \item ec.
  \item KO.
  \item GO.
  \item Best.hit.arabi.name
  \item arabi.symbol.
  \item arabi.defline.
 }
}
\references{
\url{data_blah.com}
}
\author{
Alice MacQueen \email{alice.macqueen@utexas.edu}
}
\keyword{data}
