% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbn_bigsnp2mashr.R
\name{cdbn_bigsnp2mashr}
\alias{cdbn_bigsnp2mashr}
\title{Convert bigsnpr output to mashr input dataframes.}
\usage{
cdbn_bigsnp2mashr(
  path = ".",
  snp = NULL,
  gwas_rds = NA,
  phenotypes = NA,
  clump = TRUE,
  scale = TRUE,
  numSNPs = 1000,
  model = c("linear", "logistic"),
  saveoutput = FALSE
)
}
\arguments{
\item{path}{File path to the rds files saved from bigsnpr, a character
string. Defaults to the working directory.}

\item{snp}{The "bigSNP" object used to run the gwas; needed if clump is
TRUE. Load with bigsnpr::snp_attach().}

\item{gwas_rds}{A character vector of saved gwas rds objects from bigsnpr. If NA, all *.rds files in the path will be used.}

\item{phenotypes}{A character vector of phenotype names for the GWAS RDS
objects. Must be the same length as gwas_rds, or NA. If NA, these will be
the rds file names.}

\item{clump}{Logical. Should SNPs be clumped by LD & p-value to
standardize signal strength across different LD blocks? Default is TRUE.}

\item{scale}{Logical. Should marker effects in each condition be
scaled to fall between -1 and 1? Default is TRUE.}

\item{numSNPs}{The number of most significant SNPs selected from each GWAS.
Ideally this will give 1 million or fewer total cells in the resultant
mash dataframes. Defaults to 1000.}

\item{model}{Regression used in bigstatsr. One of "logistic" or "linear".
Default is "linear".}

\item{saveoutput}{Logical. Should the function's output also be saved to RDS
files? Default is FALSE.}
}
\value{
A list containing five data frames: the SNPs selected, the B_hat
   and S_hat matrices for the strong SNP set and for a random SNP set that
   is twice the size.
}
\description{
This function converts bigsnpr output, saved as rds files to
   the specified path, to four dataframes used in the R package mashr. It
   can clump SNPs based on LD and the maximum -log10(p-value) across all
   included GWAS. It can also set the random effect data frames to come from
   a subsample of SNPs clumped by MAF and LD.
}
\note{
To create a vector of phenotype names, use the
    \code{\link{get_results_in_folder}} function.
}
\examples{
\dontrun{cdbn_bigsnp2mashr(path = system.file("inst/extdata"), numSNPs = 20,
    model = "linear")}
\dontrun{cdbn_bigsnp2mashr(numSNPs = 10000, model = "logistic")}
\dontrun{cdbn_bisgnp2mashr(numSNPs = 20000, model = "linear", saveoutput = TRUE)}
\dontrun{phenotype_vector <- get_results_in_folder(path = system.file(
    "inst/extdata"))
    numSNPs <- 1000000 / length(phenotype_vector)^2
    cdbn_bigsnp2mashr(phenotypes = phenotype_vector, numSNPs = numSNPs,
model = "linear", saveoutput = TRUE)}

}
