% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT_PC_selection.R
\name{gapit_model_selection}
\alias{gapit_model_selection}
\title{Select the best number of PCs using BIC.}
\usage{
gapit_model_selection(path, saveoutput = FALSE)
}
\arguments{
\item{path}{File path to the csv Result files that GAPIT has created, a
character string or \code{file.path()}.}

\item{saveoutput}{Should the output of this function be saved to disc?
Recommended to be set to TRUE, but the default is FALSE.}
}
\value{
A tbl_df table of traits with the best number of PC's as determined
by the maximum BIC.
}
\description{
If you choose \code{Model.Selection = TRUE} in GAPIT, you will get
BIC.Model.Selection output files. If you have many traits, it can be tedious
to manually go through each output file and select the number of principle
components that maximizes BIC. This function automates that process for many
traits in some filepath. It returns a data frame containing the trait names
in a filepath and the best number of PC's, a.k.a. the number that maximizes
the BIC for that trait.
}
\examples{
\dontrun{gapit_model_selection(path = "./inst/extdata", saveoutput = TRUE)}
\dontrun{gapit_model_selection(path = "./inst/extdata")}

}
