% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{get_annotations}
\alias{get_annotations}
\title{Create a Table of Annotated Top SNPs for Phaseolus vulgaris}
\usage{
get_annotations(
  df,
  type = c("bigsnp", "gapit", "mash", "rqtl2", "table"),
  n = 10,
  FDRalpha = NA,
  rangevector = c(0, 10000),
  markers = NULL,
  anno_info = NULL,
  txdb = NULL
)
}
\arguments{
\item{df}{a data frame or tbl_df. Can be bigsnpr output, mashr output
(loaded into R), gapit or r/qtl2 output (specify the path to a saved
.csv file), or, for Phaseolus vulgaris intervals in another format, a
data frame containing columns 'CHR', 'start', and 'end'.}

\item{type}{Type of Phaseolus vulgaris genomic marker input specified by the
df parameter. Options are "bigsnp", "gapit", "mash", "rqtl2", and "table".
 Defaults to 'bigsnp'.}

\item{n}{An integer or integer vector The numberof most significant SNPs to
select (by p-value). Set to NA to omit this table. Default is 10.}

\item{FDRalpha}{The false discovery rate. Numeric, a number or vector of
numbers between 0 and 1. Set to NA to omit this table. Default is 0.1.}

\item{rangevector}{How far from the significant SNP should annotations be
pulled? Can be an integer or a vector of integers. Default is 0 (the SNP
itself) and a 10 kbp window around the SNP.}

\item{markers}{For data frames of type "mash" or "bigsnp", the same set of
markers (with CHR and POS columns) as your df object.}

\item{anno_info}{Gene information from
Pvulgaris_442_v2.1.annotation_info.txt}

\item{txdb}{Annotation information from Pvulgaris_442_v2.1.gene.txdb.sqlite}
}
\value{
A list containing dataframes of SNPs. If more than one dataframe is
    returned, they are named using the criteria used to select the SNPs in
    the dataframe.
}
\description{
Loads in one set of GWAS results from a bigsnp GWAS, a GAPIT
   GWAS, a mash object, RQTL2 output, or a dataframe with 'CHR',
   'start', and 'end' columns for the genome of Phaseolus_vulgaris. Then, it
   constructs SNP tables meeting different criteria using these genomic
   intervals.
}
