% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbn_kinship.R
\name{get_best_PC_df}
\alias{get_best_PC_df}
\title{Return best number of PCs in terms of lambda_GC for the CDBN.}
\usage{
get_best_PC_df(df)
}
\arguments{
\item{df}{Dataframe of phenotypes where the first column is NumPCs and
subsequent column contains lambda_GC values for some phenotype.}
}
\value{
A dataframe containing the best lambda_GC value and number of PCs
    for each phenotype in the data frame.
}
\description{
Given a dataframe created using get_lambda_GC, this function
    returns the first lambda_GC less than 1.05, or the smallest lambda_GC,
    for each column in the dataframe.
}
