% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mash_results.R
\name{get_colnames}
\alias{get_colnames}
\title{Get column names from a mash object}
\usage{
get_colnames(m)
}
\arguments{
\item{m}{An object of type mash}
}
\value{
A vector of phenotype names
}
\description{
This function extracts the column names from the local false
sign rate table of a mash object's results. This can tell you the condition
names or phenotype names used in the mash object. That can be useful for
looking at a subset of these columns, say.
}
\examples{
    \dontrun{get_colnames(m = mash_obj)}

}
