% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbn_kinship.R
\name{get_date_filename}
\alias{get_date_filename}
\title{Get current date-time in a filename-appropriate format.}
\usage{
get_date_filename()
}
\value{
A string containing the current date-time with spaces and colons
    replaced with underscores and periods, respectively.
}
\description{
Converts the current \code{Sys.time()} system time to a format
    that is acceptable to include in a filename. Changes punctuation that
    won't work in a filename.
}
