% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbn_kinship.R
\name{get_kinship}
\alias{get_kinship}
\title{Find a kinship matrix using the van Raden method.}
\usage{
get_kinship(snp, ind.row = NA, hasInbred = TRUE, saveoutput = FALSE)
}
\arguments{
\item{snp}{A "bigSNP" object; load with bigsnpr::snp_attach().}

\item{ind.row}{An integer vector of the rows (individuals) to find a
kinship matrix for. Defaults to all rows.}

\item{hasInbred}{Logical. Does the SNP file contain inbred individuals or
closely related individuals, like siblings? Default is TRUE.}

\item{saveoutput}{Logical. Should the output be saved to the working
directory?}
}
\value{
A kinship matrix with labeled rows and columns.
}
\description{
Calculate the kinship matrix using code from GAPIT and bigsnpr
    and the methods of VanRaden (2009, J. Dairy Sci. 91:4414???C4423). This
    code is a modified version of the same code in GAPIT. Note
    that this matrix cannot currently be used with the GWAS methods in
    bigsnpr; however, this matrix could be used for other analyses, such as
    variance components analyses.
}
\examples{
\dontrun{
K <- get_kinship(snp = snp, saveoutput = TRUE)
}

}
