% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbn_kinship.R
\name{get_lambda_GC}
\alias{get_lambda_GC}
\title{Return lambda_GC for different numbers of PCs for GWAS on any bigsnp file.}
\usage{
get_lambda_GC(
  df,
  type = c("linear", "logistic"),
  snp,
  covar = NA,
  ncores = 1,
  npcs = c(0:10),
  saveoutput = FALSE
)
}
\arguments{
\item{df}{Dataframe of phenotypes where the first column is Taxa and each
Taxa occurs only once in the dataframe.}

\item{type}{Character string. Type of univarate regression to run for GWAS.
Options are "linear" or "logistic".}

\item{snp}{Genomic information to include for Phaseolus vulgaris.}

\item{covar}{Covariance matrix to include in the regression. You
can generate these using \code{bigsnpr::snp_autoSVD()}.}

\item{ncores}{Number of cores to use. Default is one.}

\item{npcs}{Integer vector of principle components to use.
Defaults to c(0:10).}

\item{saveoutput}{Logical. Should output be saved as a csv to the
working directory?}
}
\value{
A dataframe containing the lambda_GC values for each number of PCs
    specified. This is also saved as a .csv file in the working directory.
}
\description{
Given a dataframe of phenotypes associated with Taxa and
    output from a PCA to control for population structure, this function will
    return a .csv file of the lambda_GC values for the GWAS upon inclusion
    of different numbers of PCs. This allows the user to choose a number of
    PCs that returns a lambda_GC close to 1, and thus ensure that they have
    done adequate correction for population structure.
}
