% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbn_kinship.R
\name{get_lambdagc}
\alias{get_lambdagc}
\title{Find lambda_GC value for non-NA p-values}
\usage{
get_lambdagc(ps, tol = 1e-08)
}
\arguments{
\item{ps}{Numeric vector of p-values. Can have NA's.}

\item{tol}{Numeric. Tolerance for optional Genomic Control coefficient.}
}
\value{
A lambda GC value (some positive number, ideally ~1)
}
\description{
Finds the lambda GC value for some vector of p-values.
}
