% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mash_results.R
\name{get_marker_df}
\alias{get_marker_df}
\title{Get mash marker_df}
\usage{
get_marker_df(m)
}
\arguments{
\item{m}{An object of type mash}
}
\description{
Pulls the names of the SNP markers from the mash object.
}
