% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbn_kinship.R
\name{get_SVD}
\alias{get_SVD}
\title{Wrapper for the snp_autoSVD function for the CDBN.}
\usage{
get_SVD(snp, k = 10, ncores = 1, saveoutput = FALSE, ...)
}
\arguments{
\item{snp}{A "bigSNP" object; load with bigsnpr::snp_attach().}

\item{k}{Integer. The number of principal components to find. Default is 10.}

\item{ncores}{Integer. Number of cores to use. Default is one.}

\item{saveoutput}{Logical. Should the output be saved to the working
directory?}

\item{...}{Other arguments to \code{\link{snp_autoSVD}}.}
}
\value{
A big_SVD object.
}
\description{
This is a wrapper to determine population structure for GWAS
    for a bigSNP object. Arguments that are recognized by
    bigsnpr::snp_autoSVD can also be specified in this function.
}
\examples{
\dontrun{
svd20 <- get_SVD(snp = snp, k = 20, saveoutput = TRUE)
}

}
