% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbn_bigsnp2mashr.R
\name{get_top_effects_log10p}
\alias{get_top_effects_log10p}
\title{Step One of bigsnp2mashr}
\usage{
get_top_effects_log10p(path, gwas_rds, phenotype, numSNPs, markers)
}
\arguments{
\item{path}{Path}

\item{gwas_rds}{RDS file with gwas results}

\item{phenotype}{Character vector. Single phenotype name}

\item{numSNPs}{Integer. Number of top SNPs to choose.}

\item{markers}{Marker CHR & POS for the GWAS you ran}
}
\description{
Step One of bigsnp2mashr
}
