% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mash_results.R
\name{get_U_by_mass}
\alias{get_U_by_mass}
\title{Get the positions of objects in a mash object Ulist that are above
     some mass threshold.}
\usage{
get_U_by_mass(m, thresh = 0.05)
}
\arguments{
\item{m}{An object of type mash}

\item{thresh}{Numeric. The mass threshold for including a covariance matrix}
}
\description{
Get the positions of objects in a mash object Ulist that are
     above some mass threshold.
}
