% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_standard_run.R
\name{load_g2m_df}
\alias{load_g2m_df}
\title{Read in the random and the strong datasets}
\usage{
load_g2m_df(path, numSNPs, suffix)
}
\arguments{
\item{path}{File path to the rds files saved from bigsnpr, a character
string. Defaults to the working directory.}

\item{numSNPs}{The number of most significant SNPs selected from each GWAS.
Ideally this will give 1 million or fewer total cells in the resultant
mash dataframes. Defaults to 1000.}

\item{suffix}{Character. Optional. Should be the unique suffix used to
save \code{cdbn_bigsnp2mashr} output as RDS files, if it was used.}
}
\value{
A list containing five data frames: the SNPs selected, the B_hat
   and S_hat matrices for the strong SNP set and for a random SNP set that
   is twice the size.
}
\description{
Read in the random and the strong datasets
}
\note{
Make sure you've removed doubles and NA's from your dataset - or set
    Shat NA's to a very large  number so that that condition will be
    discounted by mashr.
}
