% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mash_results.R
\name{mash_plot_covar}
\alias{mash_plot_covar}
\title{ggplot of covariance matrix masses}
\usage{
mash_plot_covar(m, saveoutput = FALSE, suffix = "")
}
\arguments{
\item{m}{An object of type mash}

\item{saveoutput}{Logical. Should the output be saved to the path?}

\item{suffix}{Character. Optional. A unique suffix used to save the files,
instead of the current date & time.}
}
\description{
Creates a bar plot using ggplot of the masses that are on each
    covariance matrix specified in the mash model.
}
\note{
This plot can be useful for seeing the overall patterns of effects in
    the data used in mash. Non-significant effects will add mass to the
    "no_effects" covariance matrix, while significant effects will add mass
    to one of the other covariance matrices. You can use mash_plot_Ulist()
    to plot the covariance matrix patterns themselves.
}
