% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mash_results.R
\name{mash_plot_effects}
\alias{mash_plot_effects}
\title{ggplot of single mash effect}
\usage{
mash_plot_effects(m, n = NA, i = NA, marker = TRUE, saveoutput = FALSE)
}
\arguments{
\item{m}{An object of type mash}

\item{n}{Optional. Integer or integer vector. The result number to plot, in
order of significance. 1 would be the top result, for example. Find
these with \code{\link{get_significant_results}}.}

\item{i}{Optional. Integer or integer vector. The result number to plot, in
the order of the mash object. 1 would be the first marker in the mash
object, for example. Find these with \code{\link{get_marker_df}}.}

\item{marker}{Optional. Print the marker name on the plot?}

\item{saveoutput}{Logical. Should the output be saved to the path?}
}
\description{
Creates a plot with point estimates and standard errors for
    effects of a single SNP in multiple conditions.
}
\note{
Specify only one of n or i.
}
