% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mash_results.R
\name{mash_plot_manhattan_by_condition}
\alias{mash_plot_manhattan_by_condition}
\title{Manhattan plot in ggplot colored by significant conditions}
\usage{
mash_plot_manhattan_by_condition(
  m,
  cond = NA,
  saveoutput = FALSE,
  thresh = 0.05
)
}
\arguments{
\item{m}{A mash object (outputted by mash).}

\item{cond}{A vector of phenotypes. Defaults to the names of each
column in the mash object.}

\item{saveoutput}{Logical. Should the output be saved to the path?}

\item{thresh}{Numeric. The threshold used for the local false sign rate to
call significance in a condition.}
}
\value{
A \code{tbl_df()} of the data used to make the Manhattan plot, and a
    ggplot object containing the Manhattan.
}
\description{
Takes a mash object and, for some vector of phenotypes, returns
    a Manhattan plot ggplot object (and its dataframe). Each SNP in the plot
    is colored by the number of phenotypes it is significant for. Even and
    odd chromosomes have different shapes for their SNPs, so that
    chromosome identity can be determined.
}
\examples{
\dontrun{manhattan_out <- mash_ggman_by_condition(m = m, saveoutput = TRUE)}

}
