% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mash_results.R
\name{mash_plot_sig_by_condition}
\alias{mash_plot_sig_by_condition}
\title{Significant SNPs per number of conditions}
\usage{
mash_plot_sig_by_condition(
  m,
  conditions = NA,
  saveoutput = FALSE,
  suffix = "",
  thresh = 0.05
)
}
\arguments{
\item{m}{An object of type mash}

\item{conditions}{A vector of conditions. Get these with get_colnames(m).}

\item{saveoutput}{Logical. Save plot output to a file? Default is FALSE.}

\item{suffix}{Character. Optional. A unique suffix used to save the files,
instead of the current date & time.}

\item{thresh}{What is the threshold to call an effect significant? Default
is 0.05.}
}
\value{
A list containing a dataframe of the number of SNPs significant per
    number of conditions, and a ggplot object using that dataframe.
}
\description{
For some number of columns in a mash object that correspond to
    conditions, find the number of SNPs that are significant for that number
    of conditions.
}
\examples{
  \dontrun{mash_plot_sig_by_condition(m = mash_obj, saveoutput = TRUE)}

}
