% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pv_GO}
\alias{Pv_GO}
\title{Phaseolus vulgaris gene ontology information}
\format{
A data frame with 1236881 rows and 9 variables
}
\usage{
data(Pv_GO)
}
\description{
A dataset containing gene ontology  information from v2.1 of the Phaseolus
    vulgaris genome. The variables are as follows:
}
\details{
\itemize{
  \item GENEID. Locus name in Phaseolus vulgaris.
  \item variant. Which transcriptomic variant is associated with this locus.
  \item GOCategories. GO Categories associated with this locus.
  \item GOInfo. Additional information about the gene ontology categories.
 }
}
\references{
\url{data_blah.com}
}
\author{
Alice MacQueen \email{alice.macqueen@utexas.edu}
}
\keyword{data}
