% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pv_kegg}
\alias{Pv_kegg}
\title{Phaseolus vulgaris kegg information}
\format{
A data frame with 1236881 rows and 9 variables
}
\usage{
data(Pv_kegg)
}
\description{
A dataset containing the annotation information from v2.1 of the Phaseolus
    vulgaris genome for KEGG categores (the Kyoto Encyclopedia of Genes and
    Genomes). The variables are as follows:
}
\details{
\itemize{
  \item GENEID. Locus name in Phaseolus vulgaris.
  \item variant. Which transcriptomic variant is associated with this locus.
  \item KEGG. KEGG Categories associated with this locus.
  \item KEGGInfo. Additional information about the KEGG categories.
 }
}
\references{
\url{data_blah.com}
}
\author{
Alice MacQueen \email{alice.macqueen@utexas.edu}
}
\keyword{data}
