% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mash_results.R
\name{reorder_cormat}
\alias{reorder_cormat}
\title{Reorder correlation matrix}
\usage{
reorder_cormat(cormat)
}
\arguments{
\item{cormat}{A correlation matrix}
}
\description{
Reorder correlation coefficients from a matrix of things
    (including NA's) and hierarchically cluster them
}
