% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbn_kinship.R
\name{round2}
\alias{round2}
\title{Return a number rounded to some number of digits}
\usage{
round2(x, at)
}
\arguments{
\item{x}{A number or vector of numbers}

\item{at}{Numeric. Rounding factor or size of the bin to round to.}
}
\value{
A number or vector of numbers
}
\description{
Given some x, return the number rounded to some number of
    digits.
}
