% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbn_kinship.R
\name{round_xy}
\alias{round_xy}
\title{Return a dataframe binned into 2-d bins by some x and y.}
\usage{
round_xy(x, y, cl = NA, cu = NA, roundby = 0.001)
}
\arguments{
\item{x}{Numeric vector. The first vector for binning.}

\item{y}{Numeric vector. the second vector for binning}

\item{cl}{Numeric vector. Optional confidence interval for the y vector,
lower bound.}

\item{cu}{Numeric vector. Optional confidence interval for the y vector,
upper bound.}

\item{roundby}{Numeric. The amount to round the x and y vectors by for 2d
binning.}
}
\value{
A dataframe containing the 2-d binned values for x and y, and their
    confidence intervals.
}
\description{
Given a dataframe of x and y values (with some optional
    confidence intervals surrounding the y values), return only the unique
    values of x and y in some set of 2-d bins.
}
