% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theme_oeco}
\alias{theme_oeco}
\title{Oeco ggplot theme}
\format{
A list containing ggplot2 specifications
}
\usage{
data(theme_oeco)
}
\description{
A list containing the oeco ggplot2 theme for use in cdbn_standard_gwas and
    mash_plot_ ... functions.
}
\references{
\url{data_blah.com}
}
\author{
Alice MacQueen \email{alice.macqueen@utexas.edu}
}
\keyword{data}
