% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txdb}
\alias{txdb}
\title{Annotation genomic ranges for Phaseolus vulgaris}
\format{A data frame with 1236881 rows and 9 variables}
\usage{
data(txdb)
}
\description{
A sqlite object containing the genomic ranges for v2.1 of the Phaseolus
    vulgaris genome. This object can be used with the AnnotationDbi package
    and the function 'loadDb'.
}
\references{
\url{data_blah.com}
}
\author{
Alice MacQueen \email{alice.macqueen@utexas.edu}
}
\keyword{data}
