% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_standard_run.R
\name{mash_standard_run}
\alias{mash_standard_run}
\title{A standard run of mashr}
\usage{
mash_standard_run(path, list_input = NA, numSNPs = NA,
  saveoutput = FALSE, U_ed = NA)
}
\arguments{
\item{path}{File path to the rds files saved from gapit2mashr, a character
string. Defaults to the working directory.}

\item{list_input}{A list containing five data frames: the SNPs selected, the
B_hat and S_hat matrices for the strong SNP set and for a random SNP set
chosen in gapit2mashr.}

\item{numSNPs}{The number of most significant SNPs selected from each GWAS.
Ideally this will give 1 million or fewer total cells in the resultant
mash dataframes. Defaults to 1000.}

\item{saveoutput}{Logical. Should the function's output also be saved to RDS
files? Default is FALSE.}

\item{U_ed}{An optional character vector}
}
\value{
A mash result, manipulable using functions in mashr and by mash_plot
    functions in the CDBNgenomics package.
}
\description{
If you have prepared mash input using the gapit2mashr R package,
    use this function on the output to run mash as recommended by vignettes
    in the mashr package.
}
\note{
This is a convenience function for users who have prepared their data
    using gapit2mashr. If you have not used gapit2mashr to make your mash
    input, you should not use this function - instead, follow the
    recommendations of the vignettes in the mashr package itself.
}
